put all API formatting functions in renamed APIformatting module
